Refactor and modularize components
Use recoil more, and rely on less prop drilling
