+++ frontend/src/components/Pages/YTC/Calculator/ApproveAndSimulateButton.tsx
@@ -1,8 +1,17 @@
 import { Button, ButtonProps, Spinner } from "@chakra-ui/react";
 import { BalancerApproval, ERC20Approval } from "../../../../features/approval/Approval";
-import { useRecoilValue } from 'recoil';
-import { isSimulatedSelector, isSimulatingAtom } from "../../../../recoil/simulationResults/atom";
+import { useRecoilState, useRecoilValue } from 'recoil';
+import { isSimulatedSelector, isSimulatingAtom, selectedCalculatorGainSelector, simulationResultsAtom } from "../../../../recoil/simulationResults/atom";
 import { deployments } from "../../../../constants/apy-mainnet-constants";
+import { useEffect, useState } from "react";
+import { useWeb3React } from "@web3-react/core";
+import { Web3Provider } from '@ethersproject/providers';
+import { elementAddressesAtom } from "../../../../recoil/element/atom";
+import { notificationAtom } from "../../../../recoil/notifications/atom";
+import { slippageToleranceAtom } from "../../../../recoil/transactionSettings/atom";
+import { ERC20__factory } from "../../../../hardhat/typechain";
+import { getBalance } from "../../../../features/element";
+import { executeYieldTokenCompounding } from "../../../../features/ytc/executeYieldTokenCompounding";
 
 interface ApproveAndSimulateButtonProps {
     tokenAddress: string | undefined;
@@ -29,28 +38,103 @@ export const ApproveAndSimulateButton: React.FC<ApproveAndSimulateButtonProps & 
 }
 
 interface ApproveAndConfirmButtonProps {
-    tokenAddress: string | undefined;
-    tokenName: string | undefined;
-    isLoading: boolean;
-    handleExecuteTransaction: () => void;
 }
 
 export const ApproveAndConfirmButton: React.FC<ApproveAndConfirmButtonProps & ButtonProps> = (props) => {
+    const {...rest } = props;
 
-    const { tokenAddress, tokenName, isLoading, handleExecuteTransaction, ...rest } = props;
+    const [isLoading, setIsLoading] = useState<boolean>(false);
+    const [balance, setBalance] = useState<number | undefined>(undefined);
+    const setSimulationResults = useRecoilState(simulationResultsAtom)[1];
+    const elementAddresses = useRecoilValue(elementAddressesAtom);
+    const { library, account } = useWeb3React();
+    const provider = library as Web3Provider;
+    const setNotification = useRecoilState(notificationAtom)[1];
+
+
+    const slippageTolerance = useRecoilValue(slippageToleranceAtom);
+
+    const selectedResult = useRecoilValue(selectedCalculatorGainSelector);
+
+    const sufficientBalance = selectedResult && balance && (balance >= selectedResult.inputs.amountCollateralDeposited);
+
+    // get the user balance of the token;
+    useEffect(() => {
+        if (account && selectedResult){
+            const erc20 = ERC20__factory.connect(selectedResult.inputs.baseTokenAddress, provider);
+
+            setBalance(undefined);
+            getBalance(account, erc20).then((res) => {
+                setBalance(res);
+            })
+        }
+    }, [selectedResult, account, provider])
+
+    if (!selectedResult){
+        return <></>
+    }
+
+    // Execute the actual calculation transaction
+    const handleExecuteTransaction = () => {
+        const signer = provider.getSigner(account || undefined);
+        if (!balance || balance < selectedResult?.inputs.amountCollateralDeposited){
+            setNotification((currentNotifications) => {
+                return [
+                    ...currentNotifications,
+                    {
+                        text: "YTC Failed",
+                        type: "ERROR",
+                        details: "User does not have enough tokens"
+                    }
+                ]
+            })
+        } else if (signer) {
+            setIsLoading(true);
+            executeYieldTokenCompounding(
+                selectedResult.inputs,
+                selectedResult.receivedTokens.yt.amount,
+                (selectedResult.spentTokens.baseTokens.amount),
+                slippageTolerance,
+                elementAddresses,
+                signer
+            ).then((receipt) => {
+                setSimulationResults([]);
+                setNotification((currentNotifications) => {
+                    return [
+                        ...currentNotifications,
+                        {
+                            text: "YTC Execution Succesful",
+                            type: "SUCCESS",
+                            linkText: "View on Explorer",
+                            link: `https://etherscan.io/tx/${receipt.transactionHash}`
+                        }
+                    ]
+                }
+            );
+            }).finally(() => {
+                setIsLoading(false)
+            })
+        }
+    }
+
 
         return <ERC20Approval
-            tokenAddress={tokenAddress}
-            tokenName={tokenName}
+            tokenAddress={selectedResult.inputs.baseTokenAddress}
+            tokenName={selectedResult.spentTokens.baseTokens.name}
             approvalAddress={deployments.YieldTokenCompounding}
             amount={undefined}
             {...rest}
         >
             <Button
                 onClick={handleExecuteTransaction}
+                disabled={!sufficientBalance}
                 {...rest}
             >
-                {isLoading ? <Spinner/> : "CONFIRM TRANSACTION"}
+                {isLoading ? 
+                    <Spinner/> : 
+                    (sufficientBalance ? 
+                        "CONFIRM TRANSACTION" : "INSUFFICIENT BALANCE"
+                    )}
             </Button>
         </ERC20Approval>
 }
